Move non-error related code to `TapirController` rather than `TapirErrorHelpers`
