+++ myndla-api/src/main/scala/no/ndla/myndlaapi/controller/ConfigController.scala
@@ -31,8 +31,6 @@ trait ConfigController {
 
     override protected val prefix: EndpointInput[Unit] = "myndla-api" / "v1" / serviceName
 
-    import ErrorHelpers._
-
     val pathConfigKey: EndpointInput.PathCapture[ConfigKey] =
       path[ConfigKey]("config-key")
         .description(s"The of configuration value. Can only be one of '${ConfigKey.all.mkString("', '")}'")
